Prepared auto release process for cross-compilation (#42)
This change is the last step in automatic release process.

* Unified build script for linux & macos hosts
* Automatic build of cross-platform binaries and packing them to tar and zip archives
* New release process will automatically attach artifacts to the release in GitHub
* Switched from unix to syscall in order to cover windows platform quit process
* Added doc with step-by-step instruction of deploying fish node as linux systemd service
* Updated sqlite & gopsutil to not fail on win/freebsd systems

+++ check.sh
@@ -64,7 +64,12 @@ fi
 echo
 echo '---------------------- YAML Lint ----------------------'
 echo
-docker run --rm -v "${root_dir}:/data" cytopia/yamllint:1.22 --strict cmd docs lib
-errors=$((${errors}+$?))
+if command -v docker >/dev/null; then
+    docker run --rm -v "${root_dir}:/data" cytopia/yamllint:1.22 --strict cmd docs lib
+    errors=$((${errors}+$?))
+else
+    # TODO: Find some useful yaml lint in go ecosystem
+    echo 'WARN: Skipping, no docker installed'
+fi
 
 exit ${errors}
